# Conflicts:
#	Assets/Scenes/Menu.unity
#	Assets/Scripts/EntityMovement.cs
#	Assets/Scripts/IStats.cs
#	Assets/Scripts/IStats.cs.meta
#	Assets/Scripts/Inventory.cs

+++ Assets/Scripts/IStats.cs
@@ -1,4 +1,4 @@
-public interface IStats { 
-        public void ModifyHealth(float modify); 
+public interface IStats {
+        public void ModifyHealth(byte modify); 
         void Die();
 }
